split out dataStoreKind.ts
